Updated board, Updated program
Code changes:
-S_LEFT initializing properly (Was commented out due to early prototype
code)
-Removing SD card will now properly draw the badge and won't stick on
the "Returning to badge mode" screen.
-Added support for SD power control pin. If SD_CD (Card Detect) is HIGH,
then there is no card in the slot and SD_POW is low. Once SD_CD reads
LOW and indicates there's a SD card, SD_POW is turned HIGH, and then the
card read is attempted.

PCB:
-Removed SD_CD's 10k resistor. This was an attempt to fix a problem that
was completely different. Problem is when SD card is inserted, the whole
device resets. Fairly confident this is a temporary short between VCC
and Ground, which is rectified in the code above.